requirements.txt
# Core Python packages
python-dotenv
requests
numpy
pandas
scipy
matplotlib
scikit-learn
pyarrow
sentencepiece
tiktoken

# PyTorch - adjust based on your system (CPU/CUDA) Not used now but will be
# For CPU only:
# torch --index-url https://download.pytorch.org/whl/cpu
# For CUDA 11.8, replace above with:
# torch --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1, replace with:
# torch --index-url https://download.pytorch.org/whl/cu121

# DB / pool
psycopg[binary,pool]

# Application and CLI tools
pyinstaller
rich
beautifulsoup4

# Ingestion helpers
unstructured
unstructured[pdf]
unstructured[docx]
unstructured[image]
markdown
pytesseract
pymupdf

# LLM APIs
openai
anthropic
google-generativeai
#Langchain
langchain
langchain_core
langchain-openai
langchain-community
langchain-text-splitters
langchain-postgres


#Caching
redis
langchain-redis


#Re-Ranking
cohere
langchain-cohere

#Evals
ragas

# UI and utilities
gradio


